Refactored Database
